requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
bcrypt==4.1.2
beautifulsoup4==4.14.2
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
decorator==5.2.1
defusedxml==0.8.0rc2
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-paypal==2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.15
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.116.1
flake8==7.3.0
Flask==3.1.2
fonttools==4.56.0
git-filter-repo==2.47.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-dialogflow==2.39.1
googleapis-common-protos==1.69.0
greenlet==3.2.3
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
idna==3.10
inflection==0.5.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
lxml==5.3.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
MouseInfo==0.1.3
oauthlib==3.2.2
packaging==26.0
parso==0.8.5
passlib==1.7.4
paypalrestsdk==1.13.3
pillow==11.1.0
prompt_toolkit==3.0.52
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydyf==0.11.0
pyflakes==3.4.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==2.0.1
pyOpenSSL==25.0.0
pyperclip==1.11.0
pyphen==0.17.2
PyRect==0.2.0
PyScreeze==1.0.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
python3-openid==3.2.0
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
PyYAML==6.0.3
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
social-auth-core==4.5.4
soupsieve==2.8
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.2
tinycss2==1.4.0
tinyhtml5==2.0.0
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
wcwidth==0.2.14
weasyprint==64.1
webencodings==0.5.1
Werkzeug==3.1.3
wikipedia==1.4.0
xlwt==1.3.0
zopfli==0.2.3.post1
